Conflicts:
	dep/mysqllite/mysys/my_gethostbyname.c
	src/server/authserver/Main.cpp
	src/server/authserver/Server/AuthSocket.cpp
	src/server/game/Accounts/AccountMgr.cpp
	src/server/game/Chat/Channels/Channel.cpp
	src/server/game/Chat/Commands/Level2.cpp
	src/server/game/Chat/Commands/Level3.cpp
	src/server/game/DataStores/DBCStores.cpp
	src/server/game/DataStores/DBCStructure.h
	src/server/game/Entities/Pet/Pet.cpp
	src/server/game/Entities/Player/Player.cpp
	src/server/game/Entities/Player/Player.h
	src/server/game/Entities/Unit/Unit.cpp
	src/server/game/Entities/Vehicle/Vehicle.h
	src/server/game/Server/Protocol/Handlers/ChannelHandler.cpp
	src/server/game/Server/Protocol/Handlers/CharacterHandler.cpp
	src/server/game/Server/Protocol/Handlers/MovementHandler.cpp
	src/server/game/Server/Protocol/Handlers/PetHandler.cpp
	src/server/game/Server/Protocol/Opcodes.cpp
	src/server/game/Server/Protocol/Opcodes.h
	src/server/game/Server/WorldSession.cpp
	src/server/game/Server/WorldSocket.cpp
	src/server/game/Spells/SpellEffects.cpp
	src/server/game/Spells/SpellMgr.cpp
	src/server/scripts/Commands/cs_account.cpp
	src/server/scripts/Commands/cs_debug.cpp
	src/server/scripts/EasternKingdoms/BaradinHold/boss_argaloth.cpp
	src/server/scripts/EasternKingdoms/BaradinHold/instance_baradin_hold.cpp
	src/server/scripts/EasternKingdoms/ScarletMonastery/boss_headless_horseman.cpp
	src/server/scripts/Examples/example_creature.cpp
	src/server/scripts/Kalimdor/TempleOfAhnQiraj/boss_cthun.cpp
	src/server/scripts/Northrend/Gundrak/boss_drakkari_colossus.cpp
	src/server/scripts/Northrend/Nexus/Oculus/boss_varos.cpp
	src/server/scripts/Northrend/Nexus/Oculus/instance_oculus.cpp
	src/server/scripts/Northrend/Nexus/Oculus/oculus.h
	src/server/scripts/Northrend/storm_peaks.cpp
	src/server/shared/Database/Implementation/CharacterDatabase.cpp
	src/server/shared/Database/Implementation/CharacterDatabase.h
	src/server/shared/SystemConfig.h
	src/server/worldserver/CommandLine/CliRunnable.cpp
	src/server/worldserver/TCSoap/TCSoap.cpp
	src/tools/extractor/MPQ.cpp
	src/tools/extractor/dbc.cpp

+++ src/server/game/Guilds/Guild.cpp
@@ -643,7 +643,7 @@ bool Guild::Member::LoadFromDB(Field* fields)
              fields[21].GetUInt8(),
              fields[22].GetUInt16(),
              fields[23].GetUInt32());
-    m_logoutTime    = fields[24].GetUInt64();
+    m_logoutTime    = fields[24].GetUInt32();
 
     if (!CheckStats())
         return false;
@@ -2799,7 +2799,7 @@ void Guild::_MoveItems(MoveItemData* pSrc, MoveItemData* pDest, uint32 splitedAm
             player->GetName(), player->GetGUIDLow(), pItemSrc->GetEntry(), tabId, slotId, destTabId, destSlotId, pItemSrc->GetEntry());
         //return; // Commented out for now, uncomment when it's verified that this causes a crash!!
     }
-    //*/
+    // */
 
     // 3. Check destination rights
     if (!pDest->HasStoreRights(pSrc))
